Refactored command and graph widget
